Disable RStudio startup HTTPS warning
Added Renviron file that includes an RStudio-specific
variable to disable the HTTPS warning text of

"WARNING: Your CRAN mirror is set to "file://~/cran" which has an insecure
(non-HTTPS) URL. The repository was likely specified in
.Rprofile or .Rprofile.site so if you wish to change it you may need to
edit one of those files. You should either switch to a repository
that supports HTTPS or change your RStudio options to not require HTTPS
downloads.

To learn more and/or disable this warning message see the
"Use secure download method for HTTP" option in Tools -> Global Options ->
Packages."

This variable was found at:

https://support.rstudio.com/hc/en-us/articles/206827897-Secure-Package-Downloads-for-R